requirements.txt
browser-use>=0.1.0
langchain>=0.1.0
langchain-aws>=0.1.0
boto3>=1.28.0
langgraph>=0.1.0
streamlit>=1.28.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.0
webdriver-manager>=4.0.0 